[FIX] IOHelper.move нормально работает на linux.

+++ LauncherCore/src/main/java/pro/gravit/utils/helper/IOHelper.java
@@ -602,7 +602,7 @@ public final class IOHelper {
 
         @Override
         public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-            Files.delete(dir);
+            if (!this.from.equals(dir)) Files.delete(dir);
             return FileVisitResult.CONTINUE;
         }
     }
